--- .travis.yml
@@ -1,9 +1,0 @@
-language: node_js
-node_js:
-  - "6.2.0"
-before_install:
-  - export DISPLAY=:99.0
-before_script:
-  - npm install
-  - npm install phantomjs-prebuilt
-script: node_modules/karma/bin/karma start test/karma-travis.js
