+++ test/SpecRunner.html
@@ -10,8 +10,7 @@
     <script type="text/javascript" src="libs/jasmine-2.4.1/jasmine.js"></script>
     <script type="text/javascript" src="libs/jasmine-2.4.1/jasmine-html.js"></script>
     <script type="text/javascript" src="libs/jasmine-2.4.1/boot.js"></script>
-    <script type="text/javascript" src="../src/vdom.js"></script>
-    <script type="text/javascript" src="specs/vDOMSpec.js"></script>
-    <script type="text/javascript" src="specs/DOMSpec.js"></script>
+    <script type="text/javascript" src="../src/main.js"></script>
+    <script type="text/javascript" src="specs/docSpec.js"></script>
   </body>
 </html>
