--- test/x-test.html
@@ -1,36 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <title>Manual test</title>
-  </head>
-  <body xmlns="http://www.w3.org/1999/xhtml">
-    <div id="hello"></div>
-    <script type="text/javascript" src="../src/vdom.js"></script>
-    <script type="text/javascript">
-      DOM.ready(() => {
-        var a = vDOM.create('DIV', {type: 'text'});
-        a.setAttribute('title', 'Hello');
-        let ul = a.append('ul');
-        ul.setAttribute('class', 'nano');
-        for (var i = 0; i < 5; i++){
-          ul.append('li', {id: 'li_' + i, text: 'This is li ' + i});
-        }
-
-        a.render('hello');
-
-        ul.append('li', {text: 'New item'});
-
-        var x1 = vDOM.get(a.tagId, true);
-        console.log(x1);
-        setTimeout(() => {
-          var x2 = vDOM.get(a.tagId, true);
-          console.log('x2.virtual');
-          console.log(x2.virtual);
-          console.log('x2.mirror');
-          console.log(x2.mirror);
-        }, 1000)
-      });
-    </script>
-  </body>
-</html>
